Probably violating Nim here: string interning.
To optimize the speed of execution, like in C++, I "intern" the strings.
Basically, when loading the words, I store them in a global container
("allWords") - and in the used words container that is mutated during
the recursion, I maintain the *pointers* to the words, instead of
the words themselves.  This means that when checking whether the
container has seen a word before or not, the comparison compares
pointers - not strings and their contents.

Much faster!

But if a GC happens, I guess we're hosed - ergo, why I believe I am
violating Nim's memory rules by doing this :-)

Am I?